requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
discord.py==1.0.0a1632+gb79a53c
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
oauthlib==2.1.0
Pillow==6.2.2
pycparser==2.19
PyNaCl==1.2.1
PySocks==1.6.8
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.2
websockets==6.0
yarl==1.3.0
